Add -Wall/-Werror and clean up code (#3)
* Add -Wall/-Werror and clean up code

Add pedantic build options to encourage cleaner code. This also removes
some of the manual sanity checking of integers in the argparse code in
favor of using the response of stroul to check if it found any errors.

Tested: builds and runs as before.


* Add the MSVC compile option and strings.h

The compile options for MSVC were not set.
Building on some machines requires strings.h as well.

+++ args.h
@@ -15,6 +15,7 @@
 */
 #pragma once
 #include <stdint.h>
+#include <stddef.h>
 #define EXPECT_VERSION 1
 
 #define CANCELLATION_BIT 0x100
@@ -121,7 +122,7 @@ typedef struct _ARGUMENTS
     uint8_t version;
     uint8_t verbose;
     uint8_t parse;
-    uint32_t align;
+    size_t align;
     uint32_t blockpad;
     char *inputBinary;
     char *outputBinary;
